feat: rename and added should pan threshold
