+++ react/components/organisms/index.js
@@ -1,4 +1,4 @@
 export * from "./button-group";
 export * from "./chat";
 export * from "./image-list";
-export * from "./tabs";
+export * from "./tabs-swipeable";
