+++ react/components/organisms/tabs-swipeable/index.js
@@ -1,0 +1,1 @@
+export * from "./tabs-swipeable";
