--- react/components/organisms/tabs/index.js
@@ -1,1 +1,0 @@
-export * from "./tabs";
